refactor app.py
